Разбить на модули: main.js: компоненты вынесены в src/components/
