+++ src/components/load-more-button.js
@@ -1,0 +1,10 @@
+const getLoadMoreButtonComponent = () => {
+  return `
+      <button class="load-more" type="button">load more</button>
+    `;
+};
+
+
+export {
+  getLoadMoreButtonComponent
+};
